Add files via upload
Abstract – The price of a new car in the industry is fixed by
the manufacturer with some additional costs incurred by the
Government in the form of taxes. So, customers buying a new
car can be assured of the money they invest to be worthy. But,
due to the increased prices of new cars and the financial
incapability of the customers to buy them, Used Car sales are
on a global increase. Therefore, there is an urgent need for a
Used Car Price Prediction system which effectively determines
the worthiness of the car using a variety of features. Existing
System includes a process where a seller decides a price
randomly and buyer has no idea about the car and it’s value in
the present day scenario. In fact, seller also has no idea about
the car’s existing value or the price he should be selling the car
at. To overcome this problem we have developed a model
which will be highly effective. Regression Algorithms are used
because they provide us with continuous value as an output
and not a categorized value. Because of which it will be
possible to predict the actual price a car rather than the price
range of a car. User Interface has also been developed which
acquires input from any user and displays the Price of a car
according to user’s inputs.